+++ src/components/AutocompleteListboxWithExpansion.tsx
