+++ src/pages/search/advanced_search/JournalFilter.tsx
@@ -8,7 +8,7 @@ import { ResultParam } from '../../../api/searchApi';
 import {
   AutocompleteListboxWithExpansion,
   AutocompleteListboxWithExpansionProps,
-} from '../../../components/AutocompletePaper';
+} from '../../../components/AutocompleteListboxWithExpansion';
 import { AutocompleteTextField } from '../../../components/AutocompleteTextField';
 import { Journal } from '../../../types/registration.types';
 import { dataTestId } from '../../../utils/dataTestIds';
